This patch adds websocket support to ndt.
It also adds a safe integer arithmetic library in third_party/safe_iop.h, and
it adds a unit testing framework in unit_testing.c.  You can see the unit tests
get run via "make check".

This adds websocket capabilities to the server by having it sniff the first
incoming message to the control socket.  If the first incoming messate is GET,
then we assume that the socket is a websocket and try to perform the websocket
handshake.  otherwise, everything proceeds as normal.

Although this test required lots of changes to lots of places, no code changes
were required for the web100clt, and all existing clients should work as
before.